requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
praw>=7.7.0
twikit>=2.0.0
rich>=13.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
supabase>=2.0.0
